# RocketTE(WebEdition)/bzwapps/app/bzshared/doc/swagger/bzshared.yaml
swagger: "2.0"
info:
  description: "API document for Rocket BlueZone Web"
  version: "1.0.0"
  title: "Rocket BlueZone Shared APIs"
  contact:
    email: "jgao@rocketsoftware.com"
  license:
    name: "Rocket Software Commercial License"
host: "rs28:30004"
basePath: "/ZLUX/plugins/com.rs.bzshared/services"
...
schemes:
- "https"
- "http"
paths:
...
  /launchpad:
    get:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "launchpad"
      operationId: "getLaunchpad"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
    delete:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "launchpad"
      operationId: "deleteLaunchpad"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
    put:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "launchpad"
      operationId: "createLaunchpad"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
  /preference:
    get:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "preference"
      operationId: "getPreference"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
    delete:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "preference"
      operationId: "deletePreference"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
    put:
      tags:
      - "userResource"
      summary: "Rocket BlueZone Web API"
      description: "preference"
      operationId: "createPreference"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "success"
        "400":
          description: "Bad Request"
...